Refactoring
I can't believe it was this easy all along. I just needed to LITERALLY
write (keymap). All this time, I thought the documentation for which-key was
using (keymap) as a placeholder for a map that I myself am supposed to
provide. NOPE. Just a magical sequence of characters. I should be angry that the
solution was right there in front of me, but honestly I'm just happy I finally
figured out what I was doing wrong after all these hours.