Refactored most of the keybinds
I'm happier with the state of these keybinds than I have been in years. I'm
debating deprecating rbon-define-key. Imagine that. Deprecating the jewel of my
config, because I've finally found a good way to do what it does using existing
functions. I'm not sure if I should be proud or not.